feat: move communicator to separate folder
