+++ tests/communicator/test_communicator.py
@@ -4,7 +4,7 @@ import pytest
 import traci
 
 from src.component import Component
-from src.wrapper.communicator import Communicator
+from src.communicator.communicator import Communicator
 
 
 @pytest.fixture
